ref(compiler): extract PhpBlockAnalyzer from DefStructSymbol
Move the :php-block parsing and __invoke arity guard out of
DefStructSymbol into a dedicated PhpBlockAnalyzer collaborator, mirroring
the injected MethodBodyAnalyzer. Keeps DefStructSymbol focused on the
defstruct special form and makes the magic-method handling independently
testable. Replaces the marker/invoke/variadic magic strings with class
constants. Behaviour unchanged.
