+++ src/php/Compiler/CLAUDE.md
@@ -79,7 +79,7 @@ Analyzer tracks param and return types via `ParamTypeInferrer`, `ReturnTypeInfer
 
 `DefStructEmitter` and `DefInterfaceEmitter` read per-symbol metadata to enrich the PHP they generate, sharing `PhpAttributeEmitterTrait` (tag + `:php/attr` reading) and the pure `Phel\Shared\PhpAttributeRenderer`. A `:tag` value can be a bare symbol/string (verbatim, so `?int`/`self`/`\DateTime` pass through), a list (union → `a|b`), or a vector (intersection → `a&b`):
 
-- `defstruct`: a field's `^{:tag <type>}` emits a typed property (`protected int $id;`); `^{:php/attr [...]}` on the struct name (class-level) or a field (property-level) emits PHP 8 attributes. `^{:php/json true}` on the struct name implements `\JsonSerializable` (emitting a `jsonSerialize()` that returns the field map) and `^{:php/stringable true}` declares `\Stringable`. A `:php` marker in the impl list opens a block of bare PHP methods (magic methods like `__invoke`/`__toString`/`__get`) emitted directly on the class with **no backing interface** — `DefStructSymbol` skips interface reflection for these and carries them as a `DefStructInterface` with an empty name, which `DefStructEmitter::interfaceNames()` filters out of the `implements` list. A custom `__invoke` must stay signature-compatible with the inherited map-lookup `__invoke(mixed $key)` (one required call arg or variadic); `DefStructSymbol::assertCompatibleInvoke()` rejects bad arity with a Phel error.
+- `defstruct`: a field's `^{:tag <type>}` emits a typed property (`protected int $id;`); `^{:php/attr [...]}` on the struct name (class-level) or a field (property-level) emits PHP 8 attributes. `^{:php/json true}` on the struct name implements `\JsonSerializable` (emitting a `jsonSerialize()` that returns the field map) and `^{:php/stringable true}` declares `\Stringable`. A `:php` marker opens a block of bare PHP magic methods (`__invoke`/`__toString`/`__get`) emitted on the class with no interface (`PhpBlockAnalyzer`, carried as a `DefStructInterface` with empty name that `DefStructEmitter` drops from `implements`); a custom `__invoke` must be 1-arg or variadic (`PhpBlockAnalyzer` rejects bad arity).
 - `definterface`: a method's arg `^{:tag <type>}` emits a typed param and the method name's `:tag` the return type; `^{:php/attr [...]}` on the interface name, a method, or a method **parameter** emits interface-/method-/parameter-level attributes (the parameter form is emitted inline: `show(#[\Autowire] string $repo)`).
 - `defenum` (`defenum*`, `DefEnumSymbol`/`DefEnumNode`/`DefEnumEmitter`): emits a native PHP `enum`; cases are keyword-named with an optional `int`/`string` value (all-or-none → backed vs pure enum), and `^{:php/attr [...]}` on the enum name emits class-level attributes. Guarded by `enum_exists`.
