+++ tests/php/Unit/Compiler/Analyzer/SpecialForm/DefStructSymbolTest.php
@@ -12,6 +12,7 @@ use Phel\Compiler\Domain\Analyzer\Environment\GlobalEnvironment;
 use Phel\Compiler\Domain\Analyzer\Environment\NodeEnvironment;
 use Phel\Compiler\Domain\Analyzer\TypeAnalyzer\SpecialForm\DefStructSymbol;
 use Phel\Compiler\Domain\Analyzer\TypeAnalyzer\SpecialForm\MethodBodyAnalyzer;
+use Phel\Compiler\Domain\Analyzer\TypeAnalyzer\SpecialForm\PhpBlockAnalyzer;
 use Phel\Lang\Symbol;
 use Phel\Shared\Exceptions\AbstractLocatedException;
 use Phel\Shared\Munge;
@@ -112,6 +113,7 @@ final class DefStructSymbolTest extends TestCase
             $this->analyzer,
             new Munge(),
             new MethodBodyAnalyzer($this->analyzer),
+            new PhpBlockAnalyzer(new Munge(), new MethodBodyAnalyzer($this->analyzer)),
         );
     }
 }
